extra_requirements.txt
# ============================================================
# Extra Requirements for distro-train
# Federated Learning P2P Platform
# ============================================================
#
# This file contains all Python dependencies needed for the project.
# Install with: pip install -r extra_requirements.txt
#
# NOTE: Some dependencies are also listed in pyproject.toml.
# This file serves as a comprehensive reference for all modules.
# ============================================================

# ============================================================
# P2P Networking & Communication
# ============================================================
multiaddr>=0.0.9
base58>=2.1.1

# libp2p - Custom fork for federated learning
# Install with: pip install git+https://github.com/lla-dane/py-libp2p.git@fed-learn
libp2p @ git+https://github.com/lla-dane/py-libp2p.git@fed-learn

# ============================================================
# Blockchain & Hedera SDK
# ============================================================
# Hedera SDK (note: may be installed as local package in p2p/hiero_sdk_python)
# hedera-sdk-py  # Uncomment if using PyPI version
protobuf>=5.28.0
grpcio==1.68.1
grpcio-tools==1.68.1

# ============================================================
# Cryptography & Security
# ============================================================
cryptography>=44.0.0
pycryptodome==3.23.0
eth-abi>=5.2.0       # Required for contract verification scripts
eth-keys==0.5.1
rlp>=4.0.0

# ============================================================
# Async & Concurrency
# ============================================================
trio>=0.25.0
trio-util

# ============================================================
# Web Framework & API Server
# ============================================================
quart-trio
hypercorn[trio]

# ============================================================
# HTTP & Networking
# ============================================================
requests>=2.32.3     # Required for mirror node API access

# ============================================================
# Machine Learning & Data Science
# ============================================================
pandas>=2.0.0
scikit-learn>=1.3.0
numpy  # Dependency of pandas/scikit-learn

# ============================================================
# Storage & IPFS
# ============================================================
pinata==0.0.1

# ============================================================
# Configuration & Environment
# ============================================================
python-dotenv>=1.0.1

# ============================================================
# CLI & Terminal UI
# ============================================================
prompt_toolkit>=3.0.0
rich>=13.0.0

# ============================================================
# Logging & Debugging
# ============================================================
# rich (already listed above for logging.RichHandler)

# ============================================================
# Development & Testing
# ============================================================
pytest>=7.0.0
black
isort
ruff

# ============================================================
# Standard Library Modules (No Installation Required)
# ============================================================
# These are built-in to Python and listed here for documentation:
# - ast
# - base64
# - json
# - os
# - random
# - time
# - datetime
# - pathlib
# - hashlib
# - subprocess
# - tempfile
# - sys
# - logging
# - collections
# - http.server
# - urllib.parse
# - dataclasses
# - typing
# - secrets
# - re

# ============================================================
# Version Information
# ============================================================
# Python Version: >=3.11, <=3.14
# Last Updated: 2025
# Project: distro-train (Federated Learning P2P Platform)
# ============================================================
